# Improve loading UX on some places in Coach