--- a/packages/kolibri-core-for-export/package.json
+++ b/packages/kolibri-core-for-export/package.json
@@ -24,7 +24,7 @@
     "js-cookie": "^3.0.5",
     "knuth-shuffle-seeded": "^1.0.6",
     "kolibri-constants": "0.2.0",
-    "kolibri-design-system": "https://github.com/MisRob/kolibri-design-system#cfd99deadd301aae83598afb5c6290085f8351fb",
+    "kolibri-design-system": "https://github.com/learningequality/kolibri-design-system#3ab9801a442b20ddf63248b17ca0d83fe21133fb",
     "lockr": "0.8.5",
     "lodash": "^4.17.21",
     "loglevel": "^1.8.1",