## Summary

- Updates Kolibri to the KDS version https://github.com/learningequality/kolibri-design-system/pull/448/ (temporary fork reference, will be replaced) to include work done in support for this PR
-  Fixes and improvements of toggling visibility in the lessons list in Coach > Plan:
    - NaN B in size column doesn't show anymore
      - By re-ordering async events
     - Addresses unresponsive toggle switch on slow networks and prevents from repeated clicks
       - By showing a loader and re-ordering async events to achieve consistency

Last item affects UX on faster networks as well. To prevent from jarring UX, a loader will be displayed for at least 2 seconds even when all requests have finished. Hopefully this helps overall to make UX feel trustworthy taking into account that toggling a lesson visibility is an important operation that would ideally give coaches some feedback.

| | Before | After |
|--|--|--|
| Slow network | ![slow-network-before](https://github.com/learningequality/kolibri/assets/13509191/93ddd0b8-74a3-46db-b543-e4f64f16ed53) | ![slow-network-after](https://github.com/learningequality/kolibri/assets/13509191/8649fa3e-52de-4cd5-96a6-a9bb5e342289) |
| Fast network | ![fast-network-before](https://github.com/learningequality/kolibri/assets/13509191/e19bd006-e66e-4648-84dd-dd89d57addf9) | ![fast-network-after](https://github.com/learningequality/kolibri/assets/13509191/4bb8d91b-b543-4f7a-8e84-6f029b2e04a9) |

## References
<!--
 * references to related issues and PRs
 * links to mockups or specs for new features
 * links to the diffs for any dependency updates, e.g. in iceqube or the perseus plugin
-->

Fixes https://github.com/learningequality/kolibri/issues/10797

## Reviewer guidance

Toggle lessons visibility in Plan > Coach on fast network as well as slow network (you can use network throttling in browser dev tools)

----

## Testing checklist

- [x] Contributor has fully tested the PR manually
- [x] If there are any front-end changes, before/after screenshots are included
- [ ] Critical user journeys are covered by Gherkin stories
- [ ] Critical and brittle code paths are covered by unit tests


## PR process

- [x] PR has the correct target branch and milestone
- [x] PR has 'needs review' or 'work-in-progress' label
- [x] If PR is ready for review, a reviewer has been added. (Don't use 'Assignees')
- [ ] If this is an important user-facing change, PR or related issue has a 'changelog' label
- [ ] If this includes an internal dependency change, a link to the diff is provided

## Reviewer checklist

- Automated test coverage is satisfactory
- PR is fully functional
- PR has been tested for [accessibility regressions](http://kolibri-dev.readthedocs.io/en/develop/manual_testing.html#accessibility-a11y-testing)
- External dependency files were updated if necessary (`yarn` and `pip`)
- Documentation is updated
- Contributor is in AUTHORS.md